Fix Firefox 76 missing tile issue
With Firefox version 76, the webrtc library issues a
trackEvent for any SDP m line in the initial offer/answer
and for any m line that had any change at all for follow
on offer/answers.  Add logic to ignore inactive m lines
or m lines that we already have a remote video tile for.