Fix duplicate tile issue due to race condition
Client code was receiving multiple track events for a remote video
before receiving the external id.  The video tile is not bound
to an attendee until the external id is received.  The result
is creating multiple tiles for a attendee.

Add external id parsing from the INDEX message which is guaranteed
to be one of the first messages received upon joining and a far client
sending video.