Fix black tiles after peerConnection reconnect
When the peer connection reconnects, the existing webrtc
tracks are ended, but the associated video tiles were not being
removed.  This caused a conflict when the new track was added and
an existing tile was already found for the remove video.

Fix is to close the remote video tile when the track ends.