Ensure video input connected before configuring in SDP
The SDP Offer option offerToReceiveVideo was being set to true if
the local video tile existed.  There is a small race condition where
the local video tile could exist but the associated transceiver input
hasn't been created yet.  The resulting behavior was a track event would
be received without a MediaStream which was throwing an error.

Change the setting of the offerToReceiveVideo flag based on the state
of the transceiver instead of the local tile.  Also add a check in the
track event to verify a stream exists and print a warning otherwise.