Fix video tiles with wrong associated attendeeID
In cases where at the same time one remote video is being unsubscribed
and another is being subscribed, the same transceiver was being used.
The result was that the same tile was being reused with no notificaton
of the change.
The fix is to not re-use the transceiver in this situation and instead
use a different one.