Reintroduce defined-in links to GitHub.
This commit breaks out doc generation into a Ruby script so we can
determine the remote at runtime. This is necessary to support forks
(which need to use 'upstream') and Travis (which uses 'origin').

This change convinces TypeDoc that this is a GitHub repo, which brings
back the "Defined in" links that I broke in my last commit.