Introduce a split between intrinsic devices and AudioTransformDevice.
This allows us to define a class of 'devices' that are not simple media
streams or device IDs, but represent a transformation of another
intrinsic device.

In this implementation, two kinds of transformation are possible,
independently or simultaneously, by implementing the
`AudioTransformDevice` interface:

* The alteration or substitution of a provided inner device at the point
  of resolution by the device controller. This can be used to rewrite
  constraints, replace one identifier with another, or to return a
  different audio input altogether, e.g., by returning a media stream.

* The insertion of a Web Audio `AudioNode` in the audio pipeline between
  the source (the inner device) and the destination, created using the
  device controller's provided `AudioContext`. This requires that
  `enableWebAudio` is called and that the `TransformDevice` return a
  (actually two nodes to allow for an arbitrary graph to be returned)
  when requested. This can be used to introduce arbitrary signal
  processing: e.g., adding a `GainNode` to alter the volume of input, or
  a `WaveShaperNode` to add reverb.

This commit also improves the barrelizer to use modern string
interpolation, and extends it with a little hack to allow the standalone
type guard function `isAudioTransformDevice` to be exported.

Finally, the `meetingV2` demo has been extended to use the
`AudioTransformDevice` interface to add optional delay to the input
audio.

Note that the name `Device` has been preserved: customer code that looks
like

```
const device: Device = …;
controller.chooseAudioInputDevice(device);
```

will continue to build with no changes required, and can be trivially
extended to

```
const device: AudioInputDevice = …;
controller.chooseAudioInputDevice(device);
```

as the builder adopts the use of `AudioTransformDevice`s.

As part of this work, `enableWebAudio` has been removed in favor of a
constructor argument to `DefaultDeviceController`. See the version
migration docs for more information.