Extend Logger.debug to support strings, not just functions.
I ran into this sharp edge, which manifested as some of my code bailing
out with `debugFunction is not a function`.

This change extends the interface to accept strings as well as
functions.

Additionally, this change adds trivial memoization in `MultiLogger`,
ensuring that debug functions are called exactly once, no matter how
many loggers will consume the resultant string.