Fix a video flicker issue on simulcast stream change
Depending on timing of messages from the server, a simulcast
stream change may trigger a video track ended and a new video
track added.  This will result in a tile remove/add.

Fix is to check both current index and index from subscribe
to see if stream being requested is from the same sender.