Fix build output to not include a nested src/.
I don't know how this ever worked, but the absence of a `rootDir`
in the `tsconfig.json` means that our attempt to target both the guides
and the `src` created nested output.

This commit introduces a second `--noEmit` config file to type-check the
guides, adds a `rootDir` to the main source config, and adjusts the doc
generator to use the broader with-docs config file for doc generation.

Output docs are unchanged, but the build output looks much more healthy.