Reselect devices in the demo app.
This covers these conditions:

* If the in-use device gets unplugged, we will select another, keeping
  Amazon Voice Focus applied if it was selected.
* If the user chose 'default' in Chrome, and the default device was
  unplugged, Chrome will select another default. This will seamlessly
  reselect 'default' and use that device.
* If the user chose 'default' in Chrome, and changed the default in the
  OS control panel, this will select the new device when that occurs..

Note that if the device list changes, and the default is selected, we
actually don't know whether the browser reassigned 'default' to a
completely different device. And our existing device stream didn't end,
so we can't use the 'ended' event. The only thing we can do is re-choose
the device.

If we can't find our previous device, then again all we can do is
re-choose default.