Generate less garbage in DefaultRealtimeController. (#889)
This eliminates the 'wrap' function, which was generating about 20 closures
per second.

Extensive testing is added to exercise each failure case. These failures
should be impossible in practice, but there we are.

Additionally, this commit removes the prohibition against the use of
stubs. Stubs are already prevalent in the codebase, and using sinon's
improves the understandability of tests.

Co-authored-by: Richard Newman <rnewman@twinql.com>