Add fatal error catching to the demo.
This change introduces the Red Screen of Death, which will be shown if
an unexpected error is thrown during execution, even if that error would
otherwise have been caught by code in the demo app.

This is largely manual: each time you add a catch block that *should
never be hit*, add a call to `fatal(err)`. Catch blocks that catch
routine failures (e.g., the browser not supporting Voice Focus) should
not add a call to `fatal`.

However, this change also adds a top-level unhandled promise rejection
listener, and a top-level error listener. These will cause fatal errors
to be detected in e.g., async tasks.

You can test this with something like:

```
setTimeout(() => Promise.reject(new Error('oh no')), 2000)
```

This change should be sufficient to alert developers and integration
tests to new programming errors.

This commit adds the following URL query params to the demo:

* `testfatal=1`: if provided, this will show the fatal error page.
* `fatal=0`: if provided, fatal errors will be ignored, even for
  localhost.
* `fatal=1`: if provided, fatal errors will cause a Red Screen of Death,
  even when not testing locally.

The default is for the Red Screen of Death to be shown if loading from
`127.0.0.1:8080` only.