Mute video elements correctly when binding.
There are two ways the SDK allows content to be shared:

* From the browser via `getDisplayMedia`, for a tab or the whole screen.
* From a `video` element (`content-share-video` in the demo), either
  from the web or from a local file.

Unfortunately, if audio is played in the content share stream, it will
be echoed back in the local conference via the main video element.

This is because the code to mute video elements on bind was setting the
`muted` _attribute_, which reflects to `defaultMuted` on the instance
itself.

This commit does the following:

* Correctly uses `element.muted = true` to set mute state.
* Sets `playinline` on the hidden video element in the demo app, for
  completeness.

Additionally, this commit refactors the demo app to correctly set the
sink ID on the `content-share-video` element when choosing an output
device. This forces non-tab content to play audio through the same
device as the rest of the conference, which will prevent echo in some
circumstances.

This change was tested by sharing the Test Video and a tab with audio in
Chrome, changing the output device, and consuming from Firefox.