Resume a suspended audio context when selecting a transform device.
In some circumstances, most notably creating a transform device prior to the
user performing a gesture (e.g., clicking), the audio context through which
the DefaultDeviceController pipes audio will be suspended.

In this commit the device controller will attempt to resume the context if
this occurs.

An exception is present for the unusual circumstance of the context being an
OfflineAudioContext.

Tests exercise the new behavior.