Rework stopping of tracks, fix several tests.
The main change in this commit is to alter the behavior of the mock for
`MediaStreamTrack.stop`. This method in reality does _not_ trigger
`onended`/`ended` observer notifications if called by application code.
Those notifications are only signaled if the track stops due to e.g.,
being remotely closed.

Fixing this necessitates introducing a new method, `externalStop`, on a
named mock class. Call this method if you want the side effects of
observers being triggered. Call the regular `stop` method to act like
you're in a browser and call the method in the normal way.

This commit also fixes a wide variety of small test errors.