Recreate local audio context when adding a transform device.
This commit factors out `ExtendedBrowserBehavior` and uses it to decide
whether to recreate a Web Audio context when selecting a transform
device that might use an Audio Worklet. This avoids a Chromium bug:

https://bugs.chromium.org/p/chromium/issues/detail?id=1173656

This commit also reworks the browser detection tests, and adds optional
logging for packets-per-second sent over the WebRTC peer connection.