Fix memory leaks. (#1115)
This commit fixes several classes to be more careful with their
lifetimes and event listeners in order to minimize leaks.

`DefaultDeviceController.releaseMediaStream` was
rewritten for clarity and to avoid leaking callbacks on input streams
when Web Audio is enabled. This code is also extended to correctly stop
the input streams of video transform devices and remove those devices
from the active device map.

In order to test memory leaks, a 'halt' mode is added to the demo. Hold
the shift key when clicking Leave. The app will clean up everything it
ought to, and you can then measure memory leaks.

Additionally, this commit introduces a `Destroyable` interface to
explicitly deallocate objects. You can use this to more easily clean up:
when you are completely finished with objects that implement this
interface, call `destroy()` to discard those objects. This interface is
likely to be expanded in future releases.

As an additional bonus, this introduces
`AsyncScheduler.nextTick(callback)`, which is equivalent to `new
AsyncScheduler(callback).start()` but without the object
creation/GC/method call overhead.