Move configuration files into config/.
Most tooling these days is configurable: we can put their dotfiles
anywhere and refer to them by path. This makes the top level a little
cleaner and easier to scan, and puts configuration together.

This will be increasingly valuable as we add more tools to the repo:
their configuration files can live in this subdirectory.