Fix integration server to handle query params while creating meeting
Currently, some canary runs are failing with the error:
Unable to create meeting, server error: Error: Request failed with status code 403

This is due to the incorrect handling of baseURL in the integration server.
This fix handles the query params if any and accordingly generates the create meeting URL.

This also changes wait time in the AppWaitTestStep to be 1000ms from 100ms.