Parallelize tasks; enable early signaling connection establishment. (#1202)
Opening the signaling websocket is one of the biggest sources of delay
when joining a meeting: even a fast connection to a nearby region can
take over 300ms.

This commit introduces an argument to `start` to allow a 'pre-start'
behavior to occur: call `start(true)` to begin signaling, and then call
`start()` to complete peer connection setup and media negotiation.

This change also parallelizes some start tasks by using a promise-based
mechanism to express a dependency chain. For the moment, this behavior
is controlled by an internal flag on `DefaultAudioVideoController`.

A sampling of integration tests are switched to use early connect,
docstrings added for the new option, and I've also added to the guide
and FAQ.

You can test this by checking the new box on the front page of the demo:
if checked, the signaling WebSocket will be established during the
lobby/device picker experience. The only visible difference should be
that the subsequent meeting join is several hundred milliseconds faster.

Co-authored-by: Richard Newman <rnewman@twinql.com>