Instantiate event reporting in demo to pass POST logger in DEBUG mode
Currently, the meeting session is created a POST logger is DEBUG mode,
hence all debug logs will make a POST request with the POST logger.
The filtering whether the debug log is related to event reporting was
handled in serverless handler. Instead, setup the event reporter in demo
to pass the POST logger just for it and not the entire meeting session.