Ensure that Amazon Voice Focus only receives downmixed mono input.
When the SDK requests an audio input, it specifies _ideal_
single-channel input. The browser is not guaranteed to return a mono
stream, and indeed builders can supply a stereo stream from their own
audio sources.

This change ensures that the Web Audio graph will provide a
single-channel mixed input to the node, and thus the processor will
receive only a single array of samples.