Fix demo button camera toggle to retain existing screen selection
Issue: Camera option selected from Device selection page is retained in
the meeting whenever video turned off and on in JS SDK meetingV2 demo.

When we toggle the button camera we choose the video input device value
from the select box which is not in the in-meeting screen rather it is
in the preview device selection screen.

The button-camera button is on the in-meeting screen and hence since it
is selecting "video-input" (preview video input HTML select box) value
it is picking and retaining what you select on the preview screen and
not retaining in-meeting screen selection.

Fix is to use stored selected video input over using video input from
preview device selection page.