Revert "Allow stubbing join info to enable easy cross account joining from browser demo (#2511)" (#2532)
This reverts commit 9f49efe2a17bffa620362c2b56fbbc1b53066698.

This is to unblock media pipeline canary alarms. Their canary is failing
as they use the gamma SDK canary demo URL in prod canaries. This change
is good but changing the URL on canary side will need a deployment and with
limited resource availablity, I wanted to avoid it.

Media capture will follow-up with correct URL update and we can re-apply this
commit then.