Added `videofx` processor.
The `videofx` processor constitutes the implementation of the new background blur and
background replacement functionalities. The processor itself is intended to be
minimalistic,
only to implement enough functionalities to load external assets from the CDN,
namely pre/post-processing javascript library, WASM modules, and the model
binaries.

This CR is the first in a series of CRs which will include additions to the demo
app and the integrations tests.

cr: https://code.amazon.com/reviews/CR-88235139