Fix false positive sendingAudioFailed events
The fix involves evaluating the SendingAudioFailureConnectionHealthPolicy
only after a session is "connected" i.e. the policy is evaluated
only if the `audioVideoDidStart()` observer is invoked, so we only
evaluate the policy after ICE negotiation is compelte or attendeePresence
is received (if wait is configured), whichever happens later.

Additionally, unused code block for connectWithTasks has been removed.