Added the background blur and replacement 2.0 to the demo app.
With this addition, the demo app has the options for three versions of
blur 2.0 (low, medium, and high)
matching roughly the same strength as the previous implementation, and two
replacement backgrounds (a beach image and a solid color).

The integration follows the pre-existing tests updated with new constants
calculated based on the new ML model.

cr: https://code.amazon.com/reviews/CR-88954521