Save the redundant audio worker code during build time so that the worker code stays intact and is able to be loaded (#2772)
The redundant audio worker code string may vary based on the build system of different projects. Because of this, certain build systems will cause invalid worker code to be generated and the redundant audio worker will not be able to run. Now, the redundant audio worker code is saved at build time so that all clients can run the redundant audio worker with the same exact code, regardless of the client's build system.

This issue was originally found in https://github.com/aws/amazon-chime-sdk-js/issues/2771.