Make uplink loss percent estimation for redundant audio more accurate (#2800)
During canary testing, it was discovered that the local uplink loss metric that was being calculated was 10-20% higher
than the actual uplink loss being simulated. Because of this discrepancy, RED kept turning off when 60% uplink loss was
being simulated because the client-side RED thought that the uplink packet loss was 65-80%, which would constantly
trigger the functionality to disable RED to avoid congestion collapse. This causes RED to turn off prematurely even when
the uplink loss is considerably lower than 75%. This change makes the estimation more accurate so that RED does not turn
off prematurely.