Setup passthrough streams for insertable streams case in the redundant audio worker (#2829)
This is so that passthrough streams do not get blocked on the main thread. Video streams were being
setup as passthrough streams in the main thread, which caused video send and receive to be blocked
by long tasks on the main thread. Running the passthrough streams in the redundant audio worker will
prevent these blocking issues.